Extract the person search logic into a component
This will be reused on the mandatory creation page.
